requirements.txt
absl-py==0.1.10
bleach==1.5.0
certifi==2016.2.28
futures==3.1.1
html5lib==0.9999999
Markdown==2.6.11
numpy==1.14.0
olefile==0.44
Pillow==4.2.1
protobuf==3.5.1
six==1.11.0
tensorflow-gpu==1.5.0
tensorflow-tensorboard==1.5.0
Werkzeug==0.14.1
wincertstore==0.2
